Refactor test helpers and lock Jest
