+++ package.json
@@ -10,8 +10,8 @@
     "test:watch": "node --experimental-vm-modules node_modules/jest/bin/jest.js --watch"
   },
   "devDependencies": {
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
+    "jest": "29.7.0",
+    "jest-environment-jsdom": "29.7.0",
     "p-limit": "^3.1.0",
     "prettier": "^3.3.2"
   },
